# oas/controller-api/swagger.v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ControllerApi",
    "description": "A controller api demonstrating swagger generation",
    "version": "2.0"
  },
  "paths": {
    "/api/v2/todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Retrieves a list of todos",
        "description": "Gets all todos, optionally filtered by category",
        "operationId": "GetTodos",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/todos/{id}": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Retrieves a todo",
        "description": "Get todo by id",
        "operationId": "GetTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Todo": {
        "required": [
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}